refactor(service): move Google Places, Config and Unsplash routes to src/service
- Move Google Places and Config routes from backend/routes to backend/src/service
- Create separate route for Unsplash in backend/src/service/unsplashRoutes.js
- Update imports in index.js and places-proxy.js to reflect new paths
- Remove old files from backend/routes
